+++ CHANGELOG.md
@@ -1,4 +1,9 @@
 CHANGELOG:
 
 **Version x.x.x**
+- 11/7/2024
+- Updated ClientList.js to pull data from the database
+
+**Version x.x.x**
+- 11/6/2024
 - Added Client Search API